Retire AdvancedQuickAssistProcessor
